streamlit-app/requirements.txt
# Streamlit
streamlit==1.29.0
requests==2.31.0
plotly==5.18.0
matplotlib>=3.7.0,<4.0.0
seaborn>=0.12.0,<1.0.0
pandas==2.1.4
python-dotenv
